+++ lib/src/dart_ipc_unix.dart
@@ -3,10 +3,6 @@ import 'dart:io';
 import 'dart_ipc_platform_interface.dart';
 
 class DartIpcUnix extends DartIpcPlatform {
-  static void registerWith() {
-    DartIpcPlatform.instance = DartIpcUnix();
-  }
-
   @override
   Future<ServerSocket> bind(String path) {
     return ServerSocket.bind(